Remove EDH authentication api methods & tests
